requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.25.2
PyPDF2==3.0.1
python-docx==1.1.0
pinecone-client==3.0.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
openai==1.3.7
groq==0.4.1
python-dotenv==1.0.0
email-validator==2.1.0
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
pydantic==2.5.2
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
plotly==5.17.0
pymupdf==1.23.14
